2024-07-07 수동 풀 ver.α

+++ src/components/user/myCertificate.jsx
@@ -9,12 +9,13 @@ import { observer } from "mobx-react";
 const MyCertificate = observer(() => {
   const [certificates, setCertificates] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
+  const ITEMS_PER_PAGE = 10; // 페이지당 보이는 데이터 개수
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
   const [editData, setEditData] = useState(null);
 
-  const ITEMS_PER_PAGE = 10; // 페이지당 보이는 데이터 개수
+
 
   const fetchCertificates = async (nickname) => {
     setLoading(true);
@@ -34,6 +35,8 @@ const MyCertificate = observer(() => {
     const nickname = authStore.getNickname();
     fetchCertificates(nickname);
   }, [authStore.getNickname()]);
+
+
 
   // 전체 페이지 수 계산
   const totalPages = Math.ceil(certificates.length / ITEMS_PER_PAGE);
@@ -73,11 +76,18 @@ const MyCertificate = observer(() => {
 
   const handleDelete = async (certificateNumber) => {
     const nickname = authStore.getNickname();
-    try {
-      await axiosClient.delete('/certificates', {params: {certificateNumber: certificateNumber}});
-      fetchCertificates(nickname);
-    } catch (error) {
-      setError("Error deleting certificate", error);
+
+    const isConfirmed = window.confirm("삭제하시겠습니까?");
+    if(isConfirmed){
+      try {
+        await axiosClient.delete('/certificates', {params: {certificateNumber: certificateNumber}});
+        fetchCertificates(nickname);
+      } catch (error) {
+        setError("Error deleting certificate", error);
+      }
+    } else {
+      // 취소 클릭 시 동작
+      console.log("삭제가 취소되었습니다.");
     }
   };
